Renamed rustc_argument_parser to rustc_arguments
